requirements.txt
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
jira>=3.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
gspread>=6.1.4
google-auth>=2.35.0
gspread-dataframe>=3.3.1

